add tooltip explaining course average
