+++ tutor/resources/styles/components/scores/index.less
@@ -5,9 +5,13 @@
 
   @nav-tab-height: 50px;
   @display-as-btn-height: 40px;
-
+  
   // negates bottom padding on body to eliminate scrollbar
   margin-bottom: -50px;
+
+  .tutor-icon {
+    color: @tutor-blue-icon;
+  }
 
   &.loading-screen-animation {
     margin-top: 120px;
